requirements.txt
aws-cdk-lib==2.50.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-apigatewayv2-integrations-alpha==2.50.0a0
boto3==1.26.3